controller_application/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
obs-websocket-py==0.5.1
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0
